more move, pass tests
