+++ examples/certutil/selfsigned/main.go
