+++ examples/grpcutil/full/Makefile
@@ -4,4 +4,4 @@ init:
 	@go get -u github.com/golang/protobuf/protoc-gen-go
 
 generate:
-	@go generate github.com/blend/go-sdk/grpcutil/examples/full/...
+	@go generate github.com/blend/go-sdk/examples/grpcutil/full/...
